requirements.txt
fake_headers==1.0.2
numpy==1.25.2
pymongo==4.6.1
python-dotenv==1.0.1
scikit_learn==1.3.0
tls_client==0.2.2